+++ src/Api/V5/Contract/AdSubtypeEnum.php
@@ -8,9 +8,13 @@ namespace Biplane\YandexDirect\Api\V5\Contract;
 class AdSubtypeEnum
 {
 
+    const MOBILE_APP_AD_BUILDER_AD = 'MOBILE_APP_AD_BUILDER_AD';
+
     const MOBILE_APP_IMAGE_AD = 'MOBILE_APP_IMAGE_AD';
 
     const NONE = 'NONE';
+
+    const TEXT_AD_BUILDER_AD = 'TEXT_AD_BUILDER_AD';
 
     const TEXT_IMAGE_AD = 'TEXT_IMAGE_AD';