+++ src/Api/V5/Contract/MobileAppAdBuilderAdFieldEnum.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Biplane\YandexDirect\Api\V5\Contract;
+
+/**
+ * Auto-generated code.
+ */
+class MobileAppAdBuilderAdFieldEnum
+{
+
+    const CREATIVE = 'Creative';
+
+    const TRACKING_URL = 'TrackingUrl';
+
+
+}
+