+++ src/Api/V5/Contract/TextAdBuilderAdAdd.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Biplane\YandexDirect\Api\V5\Contract;
+
+/**
+ * Auto-generated code.
+ */
+class TextAdBuilderAdAdd extends AdBuilderAdAddBase
+{
+
+    protected $Href = null;
+
+    /**
+     * Creates a new instance of TextAdBuilderAdAdd.
+     *
+     * @return self
+     */
+    public static function create()
+    {
+        return new self();
+    }
+
+    /**
+     * Gets Href.
+     *
+     * @return string
+     */
+    public function getHref()
+    {
+        return $this->Href;
+    }
+
+    /**
+     * Sets Href.
+     *
+     * @param string $value
+     * @return self
+     */
+    public function setHref($value)
+    {
+        $this->Href = $value;
+
+        return $this;
+    }
+
+
+}
+