+++ src/Api/V5/Contract/TextAdBuilderAdFieldEnum.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Biplane\YandexDirect\Api\V5\Contract;
+
+/**
+ * Auto-generated code.
+ */
+class TextAdBuilderAdFieldEnum
+{
+
+    const CREATIVE = 'Creative';
+
+    const HREF = 'Href';
+
+
+}
+