+++ src/Api/V5/Contract/TextAdBuilderAdUpdate.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Biplane\YandexDirect\Api\V5\Contract;
+
+/**
+ * Auto-generated code.
+ */
+class TextAdBuilderAdUpdate extends AdBuilderAdUpdateBase
+{
+
+    protected $Href = null;
+
+    /**
+     * Creates a new instance of TextAdBuilderAdUpdate.
+     *
+     * @return self
+     */
+    public static function create()
+    {
+        return new self();
+    }
+
+    /**
+     * Gets Href.
+     *
+     * @return string|null
+     */
+    public function getHref()
+    {
+        return $this->Href;
+    }
+
+    /**
+     * Sets Href.
+     *
+     * @param string|null $value
+     * @return self
+     */
+    public function setHref($value = null)
+    {
+        $this->Href = $value;
+
+        return $this;
+    }
+
+
+}
+